Refactor limit reductions form and SA parameters initialization
